+++ src/common/flags.ts
@@ -31,6 +31,7 @@ export const T_GENERIC        = 'g';
 export const T_MAPPED         = 'm';
 export const T_TRUE           = '1';
 export const T_FALSE          = '0';
+export const T_CALLSITE       = 'c';
 
 export type RtSimpleType<T>   = { TΦ: T };
 export type RtVoidType        = RtSimpleType<typeof T_VOID>;