+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# vNext 
+- Added call-site reflection: Reflect on the type information at the location where your function is called
+
 # v0.2.1
 - Fixes an issue where `RtAnyRef#kind` reported `unknown` instead of `any`
 - Add ability to detect whether a return type is inferred or explicit (#16)